+++ docs/components/radio-field/index.md
@@ -28,7 +28,7 @@ The `@value` argument is required. To tie into the change event, provide `@onCha
 To set the checked state of the radio, the `@selectedValue` and `@value` must match. If these two arguments do not match, the radio will not be checked.
 
 ```hbs
-<Form::RadioField
+<Form::Fields::Radio
   @label='Label'
   @name='options'
   @value='option-1'
@@ -68,7 +68,7 @@ Consumers have direct access to the underlying [radio element](https://developer
 Provide a custom selector via `@rootTestSelector`. This test selector will be used as the value for the `data-root-field` attribute. The Field can be targeted via:
 
 ```hbs
-<Form::RadioField @label='Label' @rootTestSelector='example' />
+<Form::Fields::Radio @label='Label' @rootTestSelector='example' />
 ```
 
 ```js
@@ -91,7 +91,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-a'
     @value='option-1'
@@ -101,7 +101,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label and hint
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-b'
     @hint='with hint text'
@@ -112,7 +112,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label and isDisabled
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-c'
     @isDisabled={{true}}
@@ -123,7 +123,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label, hint, and selectedValue
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-d'
     @hint='With hint text'
@@ -135,7 +135,7 @@ Target the hint block via `data-hint`.
 ### RadioField with long label, and short error
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='This is an option that expands to multiple lines'
     @name='options-e'
     @value='option-1'
@@ -145,7 +145,7 @@ Target the hint block via `data-hint`.
 ### RadioField with long label and hint
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='This is an option that expands to multiple lines'
     @name='options-f'
     @hint='Here is helper text that overflows onto multiple lines'
@@ -156,7 +156,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label, hint, and value isDisabled
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-g'
     @hint='With hint text'
@@ -168,7 +168,7 @@ Target the hint block via `data-hint`.
 ### RadioField with label, selectedValue, and isDisabled
 
 <div class='mb-4 w-64'>
-  <Form::RadioField
+  <Form::Fields::Radio
     @label='Label'
     @name='options-h'
     @selectedValue='option-1'
