+++ docs/components/radio-group-field/index.md
@@ -19,11 +19,11 @@ Optional. Provide a string to `@hint` to render the text into the Hint section o
 Optional. Provide a string or array of strings to `@error` to render the text into the Error section of the fieldset.
 
 ```hbs
-<Form::RadioGroupField @label='Label' @name='single-error' @error='Error' />
+<Form::Fields::RadioGroup @label='Label' @name='single-error' @error='Error' />
 ```
 
 ```hbs
-<Form::RadioGroupField
+<Form::Fields::RadioGroup
   @label='Label'
   @name='multiple-errors'
   @error={{(array 'Error 1' 'Error 2')}}
@@ -40,7 +40,7 @@ To tie into the input event when a radio is clicked, provide `@onChange`. `@onCh
 A `@value` argument to the RadioGroupField is used to determine which of the radios is currently selected. When an `@onChange` event occurs, it's most common to update the `@value` argument to the newly selected value as shown in the example below.
 
 ```hbs
-<Form::RadioGroupField
+<Form::Fields::RadioGroup
   @label='Label'
   @name='options'
   @value={{this.groupValue}}
@@ -49,7 +49,7 @@ A `@value` argument to the RadioGroupField is used to determine which of the rad
 >
   <group.RadioField @label='Option 1' @value='option-1' />
   <group.RadioField @label='Option 2' @value='option-2' />
-</Form::RadioGroupField>
+</Form::Fields::RadioGroup>
 ```
 
 ```js
@@ -74,7 +74,7 @@ export default class extends Component {
 To disable the fieldset and all child radios, set the `@isDisabled` argument directly on the radio group field.
 
 ```hbs
-<Form::RadioGroupField
+<Form::Fields::RadioGroup
   @label='Label'
   @name='options'
   @value={{this.groupValue}}
@@ -84,7 +84,7 @@ To disable the fieldset and all child radios, set the `@isDisabled` argument dir
 >
   <!-- This will now be disabled as well! -->
   <group.RadioField @label='Option 1' @value='option-1' />
-</Form::RadioGroupField>
+</Form::Fields::RadioGroup>
 ```
 
 ### Individual Radios
@@ -92,7 +92,7 @@ To disable the fieldset and all child radios, set the `@isDisabled` argument dir
 To disable individual radio fields, set the `@isDisabled` argument directly on the radio field.
 
 ```hbs
-<Form::RadioGroupField
+<Form::Fields::RadioGroup
   @label='Label'
   @name='options'
   @value={{this.groupValue}}
@@ -100,7 +100,7 @@ To disable individual radio fields, set the `@isDisabled` argument directly on t
   as |group|
 >
   <group.RadioField @label='Option 1' @value='option-1' @isDisabled={{true}} />
-</Form::RadioGroupField>
+</Form::Fields::RadioGroup>
 ```
 
 ## Attributes and Modifiers
@@ -112,7 +112,7 @@ Consumers have direct access to the underlying [radio element](https://developer
 ### RadioGroupField with label
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField
+  <Form::Fields::RadioGroup
     @label='Label'
     @name='options-a'
     as |group|
@@ -121,65 +121,65 @@ Consumers have direct access to the underlying [radio element](https://developer
     <group.RadioField @label='Option 2' @value='option-2'/>
     <group.RadioField @label='Option 3' @value='option-3'/>
     <group.RadioField @label='Option 4' @value='option-4'/>
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label and hint
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @name='options-b' @hint='Select an option' as |group|>
+  <Form::Fields::RadioGroup @label='Label' @name='options-b' @hint='Select an option' as |group|>
   <group.RadioField @label='Option 1' @value='option-1' />
   <group.RadioField @label='Option 2' @value='option-2' />
   <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label and error
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @name='options-c' @error='With error' as |group|>
+  <Form::Fields::RadioGroup @label='Label' @name='options-c' @error='With error' as |group|>
     <group.RadioField @label='Option 1' @value='option-1' />
     <group.RadioField @label='Option 2' @value='option-2' />
     <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label, hint, and error
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @name='options-d' @hint='Select an option' @error='With error' as |group|>
+  <Form::Fields::RadioGroup @label='Label' @name='options-d' @hint='Select an option' @error='With error' as |group|>
     <group.RadioField @label='Option 1' @value='option-1' />
     <group.RadioField @label='Option 2' @value='option-2' />
     <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label and isDisabled
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @isDisabled={{true}} @name='options-e' as |group|>
+  <Form::Fields::RadioGroup @label='Label' @isDisabled={{true}} @name='options-e' as |group|>
     <group.RadioField @label='Option 1' @value='option-1' />
     <group.RadioField @label='Option 2' @value='option-2' />
     <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label, hint, and isDisabled
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @name='options-f' @isDisabled={{true}} @hint='With hint text' as |group|>
+  <Form::Fields::RadioGroup @label='Label' @name='options-f' @isDisabled={{true}} @hint='With hint text' as |group|>
     <group.RadioField @label='Option 1' @value='option-1' />
     <group.RadioField @label='Option 2' @value='option-2' />
     <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
 
 ### RadioGroupField with label and multiple errors
 
 <div class='mb-4 w-64'>
-  <Form::RadioGroupField @label='Label' @name='multiple-errors' @error={{(array 'With error 1' 'With error 2' 'With error 3')}} as |group|>
+  <Form::Fields::RadioGroup @label='Label' @name='multiple-errors' @error={{(array 'With error 1' 'With error 2' 'With error 3')}} as |group|>
     <group.RadioField @label='Option 1' @value='option-1' />
     <group.RadioField @label='Option 2' @value='option-2' />
     <group.RadioField @label='Option 3' @value='option-3' />
-  </Form::RadioGroupField>
+  </Form::Fields::RadioGroup>
 </div>
