+++ docs/headless-form/changeset/demo/base-demo.md
@@ -8,7 +8,7 @@
   as |form|
 >
   <form.Field @name='name' as |field|>
-    <Form::InputField
+    <Form::Fields::Input
       @label='Name'
       @error={{this.mapErrors field.rawErrors}}
       @hint='What should we call you?'
@@ -19,7 +19,7 @@
   </form.Field>
 
   <form.Field @name='email' as |field|>
-    <Form::InputField
+    <Form::Fields::Input
       @label='Email'
       @error={{this.mapErrors field.rawErrors}}
       @hint='How do we contact you?'
