+++ ember-toucan-core/src/components/form/fields/file-input.ts
@@ -2,7 +2,7 @@ import Component from '@glimmer/component';
 import { assert } from '@ember/debug';
 import { action } from '@ember/object';
 
-import type { ErrorMessage } from '../../-private/types';
+import type { ErrorMessage } from '../../../-private/types';
 
 type FileTarget = EventTarget & { files?: FileList };
 export type FileEvent = (Event | MouseEvent) & { target: FileTarget | null };
