+++ ember-toucan-core/src/components/form/file-input/delete-button.ts
@@ -21,17 +21,17 @@ export default class ToucanFormFileInputDeleteButtonComponent extends Component<
     args: ToucanFormFileInputDeleteButtonComponentSignature['Args']
   ) {
     assert(
-      'A "@file" argument is required for Form::FileInput::DeleteButton. If using the Form::FileInputField this should be provided automatically.',
+      'A "@file" argument is required for Form::FileInput::DeleteButton. If using the Form::Fields::FileInput this should be provided automatically.',
       args.file !== undefined
     );
 
     assert(
-      'A "@onDelete" argument is required for Form::FileInput::DeleteButton. If using the Form::FileInputField this should be provided automatically.',
+      'A "@onDelete" argument is required for Form::FileInput::DeleteButton. If using the Form::Fields::FileInput this should be provided automatically.',
       args.onDelete !== undefined
     );
 
     assert(
-      'A "@deleteLabel" argument is required for Form::FileInput::DeleteButton. If using the Form::FileInputField this should be provided automatically.',
+      'A "@deleteLabel" argument is required for Form::FileInput::DeleteButton. If using the Form::Fields::FileInput this should be provided automatically.',
       args.deleteLabel !== undefined
     );
