+++ ember-toucan-core/src/components/form/file-input/list-item.ts
@@ -20,17 +20,17 @@ export default class ToucanFormFileInputListComponent extends Component<ToucanFo
     args: ToucanFormFileInputListItemComponentSignature['Args']
   ) {
     assert(
-      'An "@onDelete" argument is required for Form::FileInput::List. If using Form::FileInputField, this should be provided automatically.',
+      'An "@onDelete" argument is required for Form::FileInput::List. If using Form::Fields::FileInput, this should be provided automatically.',
       args.onDelete !== undefined
     );
 
     assert(
-      'An "@file" argument is required for Form::FileInput::List. If using Form::FileInputField, this should be provided automatically.',
+      'An "@file" argument is required for Form::FileInput::List. If using Form::Fields::FileInput, this should be provided automatically.',
       args.file !== undefined
     );
 
     assert(
-      'An "@deleteLabel" argument is required for Form::FileInput::List. If using Form::FileInputField, this should be provided automatically.',
+      'An "@deleteLabel" argument is required for Form::FileInput::List. If using Form::Fields::FileInput, this should be provided automatically.',
       args.deleteLabel !== undefined
     );
     super(owner, args);
